+++ codec/protobuf/doc.go
@@ -1,3 +1,3 @@
-// Package protobuf contains codec implemenentations for protocol buffers
+// Package protobuf contains codec implementations for protocol buffers
 // types.
 package protobuf