+++ codec/stateless/doc.go
@@ -1,0 +1,3 @@
+// Package stateless contains a codec implementation for
+// [dogma.StatelessProcessRoot] values.
+package stateless