+++ CHANGELOG.md
@@ -10,6 +10,14 @@ The format is based on [Keep a Changelog], and this project adheres to
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
+## Unreleased
+
+### Added
+
+- Added `stateless.DefaultCodec`, a codec for marshaling/unmarshaling
+  `dogma.StatelessProcessRoot` values without assuming that any other codec is
+  available.
+
 ## [0.7.3] - 2023-03-27
 
 ### Fixed